+++ src/py/flwr/driver/app.py
@@ -17,12 +17,14 @@
 
 import sys
 import threading
+import time
 from logging import INFO
-from typing import Optional
+from typing import Dict, Optional
 
 from flwr.common import EventType, event
 from flwr.common.address import parse_address
 from flwr.common.logger import log
+from flwr.proto import driver_pb2
 from flwr.server.app import ServerConfig, init_defaults, run_fl
 from flwr.server.client_manager import ClientManager
 from flwr.server.history import History
@@ -30,7 +32,7 @@ from flwr.server.server import Server
 from flwr.server.strategy import Strategy
 
 from .driver import Driver
-from .driver_client_manager_utils import update_client_manager
+from .driver_client_proxy import DriverClientProxy
 
 DEFAULT_SERVER_ADDRESS_DRIVER = "[::]:9091"
 
@@ -151,3 +153,58 @@ def start_driver(  # pylint: disable=too-many-arguments, too-many-locals
     event(EventType.START_SERVER_LEAVE)
 
     return hist
+
+
+def update_client_manager(
+    driver: Driver,
+    client_manager: ClientManager,
+    lock: threading.Lock,
+) -> None:
+    """Update the nodes list in the client manager.
+
+    This function periodically communicates with the associated driver to get all
+    node_ids. Each node_id is then converted into a `DriverClientProxy` instance
+    and stored in the `registered_nodes` dictionary with node_id as key.
+
+    New nodes will be added to the ClientManager via `client_manager.register()`,
+    and dead nodes will be removed from the ClientManager via
+    `client_manager.unregister()`.
+    """
+    # Request for workload_id
+    workload_id = driver.create_workload(driver_pb2.CreateWorkloadRequest()).workload_id
+
+    # Loop until the driver is disconnected
+    registered_nodes: Dict[int, DriverClientProxy] = {}
+    while True:
+        with lock:
+            # End the while loop if the driver is disconnected
+            if driver.stub is None:
+                break
+            get_nodes_res = driver.get_nodes(
+                req=driver_pb2.GetNodesRequest(workload_id=workload_id)
+            )
+        all_node_ids = {node.node_id for node in get_nodes_res.nodes}
+        dead_nodes = set(registered_nodes).difference(all_node_ids)
+        new_nodes = all_node_ids.difference(registered_nodes)
+
+        # Unregister dead nodes
+        for node_id in dead_nodes:
+            client_proxy = registered_nodes[node_id]
+            client_manager.unregister(client_proxy)
+            del registered_nodes[node_id]
+
+        # Register new nodes
+        for node_id in new_nodes:
+            client_proxy = DriverClientProxy(
+                node_id=node_id,
+                driver=driver,
+                anonymous=False,
+                workload_id=workload_id,
+            )
+            if client_manager.register(client_proxy):
+                registered_nodes[node_id] = client_proxy
+            else:
+                raise RuntimeError("Could not register node.")
+
+        # Sleep for 3 seconds
+        time.sleep(3)
